improve the code structure and decompose
