+++ lib/features/sber_profile/Model/Card.dart
@@ -7,4 +7,23 @@ class CardOffer {
   Image imageOfCard;
 
   CardOffer(this.cardName, this.cardInfo, this.cardCost, this.imageOfCard);
+
+  static var listOfCards = [
+    CardOffer(
+        "СберПрайм",
+        "Платеж 9 июля",
+        "199 ₽ в месяц",
+        Image.asset(
+          "assets/images/sberPraimIcon.png",
+          height: 36,
+        )),
+    CardOffer(
+        "Переводы",
+        "Автопродление 21 августа",
+        "199 ₽ в месяц",
+        Image.asset(
+          "assets/images/percentFillIcon.png",
+          height: 36,
+        )),
+  ];
 }
