+++ lib/features/sber_profile/Model/CategoryOfInterest.dart
@@ -8,4 +8,14 @@ class CategoryOfInterest {
     this.name,
     this.iconOfCategory,
   );
+
+  static var listOfCategories = [
+    CategoryOfInterest("Еда", null),
+    CategoryOfInterest("Саморазвитие", null),
+    CategoryOfInterest("Технологии", null),
+    CategoryOfInterest("Дом", null),
+    CategoryOfInterest("Досуг", null),
+    CategoryOfInterest("Забота о себе", null),
+    CategoryOfInterest("Наука", null)
+  ];
 }
