+++ lib/features/sber_profile/widgets/widgets.dart
@@ -1,2 +1,4 @@
 export 'header/header_profile.dart';
-export 'main_content.dart';
+export 'content_widgets/interest_section.dart';
+export 'content_widgets/services_section.dart';
+export 'content_widgets/tarifs_section.dart';
